Use a Buffer Pool
Instead of allocating new byte arrays each time you perform a cryptographic operation (such as signing, encryption, or converting data to/from Base64), you can use a buffer pool. A buffer pool maintains a set of preallocated buffers that can be reused, reducing the need for frequent allocations. In .NET, the ArrayPool<T> class is ideal for this purpose.